Added google analytics to track user usage, also added 2 additional levels with some unlocking features, fixed a cpl minor bugs...

+++ Hanoi/GameTimerTickEvent.cs
@@ -4,10 +4,10 @@ namespace Hanoi
 {
     public class LevelTimerTickEventArgs : EventArgs
     {
-        public LevelTimerTickEventArgs(int seconds)
+        public LevelTimerTickEventArgs(long seconds)
         {
             Seconds = seconds;
         }
-        public int Seconds { get; private set; }
+        public long Seconds { get; private set; }
     }
 }
